LayoutTests/http/tests/cookies/document-cookie-set-expired-cookie-expected.txt

Tests that setting a cookie with an expiration date in the past deletes the cookie

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.cookie is ""
PASS document.cookie is "testDelete=testValue"
PASS document.cookie is ""
PASS document.cookie is ""
PASS successfullyParsed is true

TEST COMPLETE
